requirements.txt
paramiko>=3.3.1
openai>=1.3.0
requests>=2.31.0
colorama>=0.4.6
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
tabulate>=0.9.0